refactor(wizard): split _integration_configurators.py into per-vendor modules
_integration_configurators.py had grown to 1889 lines via per-vendor
accretion (~30 _configure_<vendor>() functions). Split into
surfaces/cli/wizard/configurators/, grouped by vendor family
(observability, alerting, chat_notifications) or per-vendor where no
close sibling exists. _integration_configurators.py now holds only the
_configure_selected_integrations() dispatcher.

flow.py imports DEFAULT_GITHUB_MCP_MODE/URL directly from
configurators.github instead of through the dispatcher. test_flow.py's
monkeypatch targets are updated to patch the vendor module that now
owns each validate_*/sync_env_*/upsert_integration call.

Fixes #4066
