+++ surfaces/cli/wizard/configurators/__init__.py
@@ -1,0 +1,1 @@
+"""Per-vendor configurator modules used by the wizard onboarding flow."""
